component-ized new question form
